scsi: ufs-qcom: Remove pm_qos usage
These pieces of codes affine a pm_qos before a transfer request is
called, then disables it after the UFS clock is turned off, leaving the
CPU in a shallow idle state for longer than necessary. Furthermore, this
affines the pm_qos to all cores, wasting immense amounts of power.
Remove the code entirely in order to replace it with a different
implementation.

Change-Id: Id5e4ee8e871c3974dc31af1fad28d673deda0f8c
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Signed-off-by: rk134 <rahul@aospa.co>
Signed-off-by: dodyirawan85 <40514988+dodyirawan85@users.noreply.github.com>